format-string-helper: Set bps at very start of function
On libc on arm that was compiler thumb the arg for vsnprintf (r2) was
already overwritten by the time the bp was hit. This made the
breakpoints apparently skipped over (since the overwritten address was
evaluated as an invalid address).

This can be fixed by setting a bp at "*func" instead of "func".