Fix slow settings retrieval
The way that command-specific settings were read was very slow. Every
setting read would have to filter through all settings twice.

This change avoids building the filter when possible.

Also, since it was only ever used for the name of the settings, return a
list of names instead of a dict of name:value.

This can speed up the context command by as much as 25%