Add syscall commands (#282)
* Add syscall commands

* is-syscall: Tells whether the next instruction to execute is a syscall
instruction
* syscall-args: Gets the syscall name and arguments

This commit prepares a template for the syscall-args command. Next step
is to create a syscall table file, which allows for retrieving of
SyscallEntry that contains all relevant information for a syscall.

* Change syscall_instruction to a list

It turns out that in x86 32 bit there is a sysenter fastcall instruction that
will cause gdb to skip over int 0x80. Not sure if there is such a thing
with other archs so better to just make it into a list of possible
commands.

* Add working POC of syscall table

Now only left with populating syscall table

* Remove syscall_number_register

Apparently there is already syscall_register

* Minor fixes

* Change to use current_arch class name instead of current_arch.arch as
the former is more descriptive
* Fix syscall_register of ARM and ARM64

* Add syscall_tables to gef config in gef-extras.sh

* Address style issues

* Update docs

* Fix docstrings for syscall commands

* Add mention of gef-extras req to syscall-args