gef.py: Support dumping chunks from non-main arena
The heap chunks command takes an argument which is a heap section
address. Currently this only works if the arena is the main one.

Duplicate the logic from ptmalloc dump_heap - find the arena from
the heap_info struct at the start of the heap section and find the
right offset from there to the beginning of the chunks.