Support glibc's Safe-Linking in malloc
glibc 2.32 will add support for protection of single-linked-lists
such as the Fastbins and TCache lists. This support, which was pushed
already to glibc's master branch, is referred to us "Safe-Linking".

Safe-Linking masks the "fwd" pointers, and so this commit implements
the support for unmasking ("reveal"ing) the pointers so they could be
used properly by gef.