Allow packaging and importing as Python package
Summary:
The goal of this change is to allow gef to be built and versioned as a
Python module so that users can easily install it using `pip install
gef`.

The second piece to this is that it should be possible to `import gef`
and run the initialization when desired, instead of automatically.

Test Plan:
- With a GDB compiled with Python 3.6, run `pip install .` within the
project directory and confirm that I can run the following:
```
$ gdb
(gdb) pi
>>> import gef
>>> gef.init()
>>> <ctrl-d>
gef> # YAY it works!
```
- Confirm that sourcing it as originally intended still works as well.