glibc function call argument definition (#577)
This change adds functionality to optionally parse a json file for glibc
arguments. If this setting is set and the file can be found, then GEF
will lookup imported functions in the json, and if there's a match, it
will use the number and type of args provided instead of guessing.