Fix libc version detection if libcrypto is loaded
I noticed that gef doesn't display tcache in some cases and tracked it
down to gef confusing libcrypto (which starts with "libc") with an
unknown libc.

Fix by restructuring the code, so that files with "definitely is a
libc" names are accepted unconditionally, and files with "maybe this
is a libc" names are accepted only if their contents matches
pattern_libc_ver.