Fix gdb.execute not quoting paths (#999)
As I looked into #901, I found that we often don't quote filepaths in
`gdb.execute`. This began lazy grep that tries to ID where we do this,
and fixes where we can. Noticed that we cannot use quotes in the
filename when setting up logging, so it instead restricts the logging
path to one without spaces.