Memory map provider (#1003)
This change allows architectures to provide their own memory map by
defining `maps` in the architecture. If set, then `GefMemoryManager`
will defer to it when building maps.

This can be used for e.g. architectures that are built for use with
gdbserver.

It also adds `parse_info_mem` (Renaming the old one to
`parse_monitor_info_mem`) which parses the maps from this command. The
Black Magic Probe provides memory maps of its targets via this command.