Add c++ symbol support in xinfo. (#1028)
This change adds c++ symbol support in xinfo.

Currently, when using xinfo to get the symbol information in C++
programs, we might get incomplete data

This is because the gdb_get_location_from_symbol function could not
handle space in the symbol name.

This change updates the gdb_get_location_from_symbol, so we can handle
the space. It makes `xinfo` produces expected result.