Migrate tests to RPyC (#1040)
This is a major PR, but long overdue.

GEF can now use RPyC to completely control the tests which not only make
them way faster, also allow to control directly from the `pytest` tests
themselves `gdb` and `gef` (see examples). Tests are therefore smaller,
less messy, and more accurate than when using the legacy function
`gdb_run_cmd`.

The documentation in `docs/testing.md` has been updated accordingly.

This allowed to find a few missing things in gef.py itself, which were
added but nothing major.

Note that currently the PR is not perfect as I did not re-write the
tests: I've migrated most of the code to the new API to ensure coverage
stays identical. This was done because the PR is already big enough as
it is. More accurate tests can be added in following PRs.


Co-authored-by: Grazfather <grazfather@gmail.com>