Fix recursive loop on architectures where `flags` register is a dynamic value (#1065)
## Description

For (at least) ARM and ARM64`$arch.ptrsize` is calculated dynamically from CSPR flags, and invokes `parse_address`. However, `parse_address` requires `ptrsize` to be set to determine the alignment. So we end up in an infinite loop. Which breaks almost everything.

This commit fixes it by simply replacing `to_unsigned_long` with `int` is perfectly fine for this situation, which is what the PR does.