Supporting changes for the "got-audit" command. (#1097)
## Description

This change moves the setup of the line describing a GOT entry into a
method in order to support the "got-audit" command, which is a subclass
of the GotCommand. It also uses a full-width output from readelf, and
gets the path for `nm`, both of which also support the got-audit command
as a subclass.